+++ bin/test.dart
@@ -1,0 +1,5 @@
+// import 'package:collapsible_cards/widget/collapsible_cards.dart';
+
+// void main() {
+//   print(greet("Fayis")); // Should print: Hello, Fayis!
+// }