+++ lib/collapsible_card.dart
@@ -1,0 +1,3 @@
+library collapsible_cards;
+
+export 'widgets/collapsible_card.dart';